Fixed #8: peewee exception on too large IN statements
It's an ugly hack for now untill we refactor this code.
Previously, we had a list of unique authors and ran some SQL stmt
based on a authors-long IN subselect.
If the authors list is long (like for EN), then it might be too long and fail.

In this hack, we get around the stmt but keep the authors list in python.
Obviously that's not ideal but it's easier to fix like this espectially without testing.

WARNING: untested as I have neither time nor setup too.